spark/requirements.txt
google-cloud-bigquery
google-cloud-storage
dotenv
datetime
pyarrow
pandas
gcsfs